+++ print_statement.c
@@ -84,7 +84,7 @@ void out_way(const Way& way, bool complete = true)
   else
   {
     cout<<"<way id=\""<<way.id<<"\">\n";
-    for (vector< int >::const_iterator it2(way.members.begin());
+    for (vector< uint32 >::const_iterator it2(way.members.begin());
 	 it2 != way.members.end(); ++it2)
       cout<<"  <nd ref=\""<<*it2<<"\"/>\n";
     if (complete)
@@ -220,14 +220,14 @@ void Print_Statement::execute(MYSQL* mysql, map< string, Set >& maps)
 	MYSQL_ROW row(mysql_fetch_row(result));
 	while ((row) && (row[0]))
 	{
-	  int id(atoi(row[0]));
+	  uint32 id(atoi(row[0]));
 	  while (it2->id < id)
 	  {
 	    out_way(*it2);
 	    ++it2;
 	  }
 	  out_way(*it2, false);
-	  while ((row) && (row[0]) && (it2->id == atoi(row[0])))
+	  while ((row) && (row[0]) && (it2->id == (uint32)atoi(row[0])))
 	  {
 	    if ((row[1]) && (row[2]))
 	      cout<<"  <tag k=\""<<row[1]<<"\" v=\""<<row[2]<<"\"/>\n";
